strawberry/requirements.txt
anyio==3.6.2
black==23.3.0
cassandra-driver==3.26.0
click==8.1.3
fastapi==0.95.0
geomet==0.2.1.post1
graphql-core==3.2.3
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strawberry-graphql==0.171.0
typing_extensions==4.5.0
uvicorn==0.21.1
